+++ noodles-vcf/src/record/builder.rs
@@ -280,10 +280,7 @@ impl Builder {
     /// let keys = "GT:GQ".parse()?;
     /// let genotypes = Genotypes::new(
     ///     keys,
-    ///     vec![vec![
-    ///         Some(Value::String(String::from("0|0"))),
-    ///         Some(Value::Integer(13)),
-    ///     ]],
+    ///     vec![vec![Some(Value::from("0|0")), Some(Value::from(13))]],
     /// );
     ///
     /// let record = vcf::Record::builder()